20101103 Added key frame nodes to rendered parameter curves (CWW-12)
git-svn-id: svn://localhost/Software/Java/WorldWind@1311 4cd5884b-1955-4949-bbec-e4dd67779ed6

+++ Applications/Animator/src/au/gov/ga/worldwind/animator/ui/parametereditor/ParameterCurveBounds.java
@@ -15,18 +15,18 @@ class ParameterCurveBounds
 		this.topRight = topRight;
 	}
 	
-	ParameterCurveBounds(int minFrame, int maxFrame, double minValue, double maxValue)
+	ParameterCurveBounds(double minFrame, double maxFrame, double minValue, double maxValue)
 	{
 		this.bottomLeft = new ParameterCurvePoint(minFrame, minValue);
 		this.topRight = new ParameterCurvePoint(maxFrame, maxValue);
 	}
 	
-	int getMinFrame()
+	double getMinFrame()
 	{
 		return bottomLeft.frame;
 	}
 	
-	int getMaxFrame()
+	double getMaxFrame()
 	{
 		return topRight.frame;
 	}
